+++ pallets/programs/src/mock.rs
@@ -54,6 +54,7 @@ impl system::Config for Test {
 parameter_types! {
   pub const MaxBytecodeLength: u32 = 3;
   pub const ProgramDepositPerByte: u32 = 5;
+  pub const MaxOwnedPrograms: u32 = 1;
 }
 
 parameter_types! {
@@ -80,6 +81,7 @@ impl pallet_programs::Config for Test {
     type Currency = Balances;
     type MaxBytecodeLength = MaxBytecodeLength;
     type ProgramDepositPerByte = ProgramDepositPerByte;
+    type MaxOwnedPrograms = MaxOwnedPrograms;
     type RuntimeEvent = RuntimeEvent;
     type WeightInfo = ();
 }