+++ pallets/transaction-pause/src/mock.rs
@@ -83,12 +83,14 @@ impl pallet_balances::Config for Runtime {
 parameter_types! {
   pub const MaxBytecodeLength: u32 = 3;
   pub const ProgramDepositPerByte: u32 = 5;
+  pub const MaxOwnedPrograms: u32 = 5;
 }
 
 impl pallet_programs::Config for Runtime {
     type Currency = ();
     type MaxBytecodeLength = MaxBytecodeLength;
     type ProgramDepositPerByte = ProgramDepositPerByte;
+    type MaxOwnedPrograms = MaxOwnedPrograms;
     type RuntimeEvent = RuntimeEvent;
     type WeightInfo = ();
 }