+++ CHANGELOG.md
@@ -21,6 +21,13 @@ At the moment this project **does not** adhere to
 ### Changed
 - Crate name refactor ([#561](https://github.com/entropyxyz/entropy-core/pull/561))
 
+### Breaking Changes
+
+- In [#536](https://github.com/entropyxyz/entropy-core/pull/536/), the registered struct no longer holds a program but rather a hash of a program that is set in the set_program function 
+
+### Added
+- Pointer for Programs ([#536](https://github.com/entropyxyz/entropy-core/pull/536/))
+
 ## [0.0.9](https://github.com/entropyxyz/entropy-core/compare/release/v0.0.8..release/v0.0.9) - 2023-11-30
 
 Some of the noteworthy changes related to this release are related to better integration in Web